@@ -1,39 +1,65 @@
 {
-    "name": "neilime/ansi-escapes-to-html",
-    "type": "library",
-    "description": "Convert ANSI escapes (terminal formatting/color codes) to HTML",
-    "keywords": ["ANSI", "escapes", "highlight", "html", "convert"],
-    "homepage": "http://neilime.github.io/ansi-escapes-to-html/",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Neilime",
-            "homepage": "https://github.com/neilime",
-            "role": "Developer"
-        }
-    ],
-    "support": {
-        "issues": "https://github.com/neilime/ansi-escapes-to-html/issues"
-    },
-    "minimum-stability": "stable",
-    "autoload": {
-        "psr-4": {
-            "AnsiEscapesToHtml\\": "src/AnsiEscapesToHtml"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "TestSuite\\": "tests/TestSuite"
-        }
-    },
-    "require": {
-        "php": "^7.2"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^8.5.31"
-    },    
-    "scripts": {
-        "test": "cd tests && phpunit -d max_execution_time=0 && cd ..",
-        "test:ci": "cd tests && phpunit -d max_execution_time=0 --coverage-text --coverage-clover ../build/logs/clover.xml --coverage-html ../build/coverage/ && cd .."
+  "name": "neilime/ansi-escapes-to-html",
+  "type": "library",
+  "description": "Convert ANSI escapes (terminal formatting/color codes) to HTML",
+  "keywords": [
+    "ANSI",
+    "escapes",
+    "highlight",
+    "html",
+    "convert"
+  ],
+  "homepage": "http://neilime.github.io/ansi-escapes-to-html/",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Neilime",
+      "homepage": "https://github.com/neilime",
+      "role": "Developer"
     }
+  ],
+  "support": {
+    "issues": "https://github.com/neilime/ansi-escapes-to-html/issues"
+  },
+  "minimum-stability": "stable",
+  "autoload": {
+    "psr-4": {
+      "AnsiEscapesToHtml\\": "src/AnsiEscapesToHtml"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "TestSuite\\": "tests/TestSuite"
+    }
+  },
+  "require": {
+    "php": "^7.4 || ^8"
+  },
+  "require-dev": {
+    "pcov/clobber": "^2.0",
+    "phpstan/extension-installer": "^1.2",
+    "phpstan/phpstan": "^1.2",
+    "phpstan/phpstan-phpunit": "^1.2",
+    "phpunit/phpunit": "^9.5.26",
+    "squizlabs/php_codesniffer": "^3.7"
+  },
+  "scripts": {
+    "test": "phpunit --colors --configuration tests/phpunit.xml",
+    "test:ci": "phpunit --colors --configuration tests/phpunit.xml -d pcov.enabled=1 -d max_execution_time=0 --coverage-text --coverage-clover ./build/logs/clover.xml --coverage-html ./build/coverage/",
+    "cs": "phpcs",
+    "cbf": "phpcbf",
+    "stan": "phpstan analyse --level 5 src",
+    "ci": [
+      "@cs",
+      "@stan",
+      "@test:ci"
+    ]
+  },
+  "config": {
+    "optimize-autoloader": true,
+    "sort-packages": true,
+    "allow-plugins": {
+      "phpstan/extension-installer": true
+    }
+  }
 }
